Make Input fully async
The implementation makes use of a collection of `Stream`s: one for each connected device (organized by type, so there can only be one of each type at most) and one for notifications of new devices.
Then the `Input` struct itself implements `Stream` and polls all these sub-streams in an orderly fashion.
My `Stream` implementation is probably not great, but it works for now. I may improve it later.