Refine impl and add input events
The WM now keeps track of a keyboard-focused surface and sends keyboard/button events to this surface.
Touch and stylus events are not yet implemented because determining which events to send to which tasks is more complicated.
Device presence events are also not included for now because it is unclear how those should be sent out and to which tasks.