Simplify proto IO; expand WM tiling
The new `rmox_protocol::io::Stream` wrapper simplifies both reading and writing by providing a single interface (no more `split`ing the Unix stream) and is backed by a custom `Stream::poll_next` implementation for the read side that only needs access to the inner IO when it's being polled (compared to something like `ReadExact` from tokio which retains a mutable borrow to the IO).

For WM tiling, `shell.roots` became a full tree of containers and surfaces. The maintenance logic to keep all of surfaces and tasks, shell tree, and keyboard focus path in sync is getting trickier. Maybe at some point I will find a better abstraction or a good boundary point for the struct groupings, but for now there is a lot of reaching between the structs with many methods on each requiring mutable borrows to the others.